update: code cleanup.
