+++ Assets/Scripts/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bfbe171ce80284770831ade808ba6ed2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
