+++ Assets/Scripts/Map/Collapse.cs
